+++ lib/commands/options/license_option_handler.rb
