+++ lib/commands/options/search_options_parser.rb
