+++ lib/commands/show_command_handler.rb
@@ -1,5 +1,5 @@
-require './lib/ruby_gems_client'
-require './lib/ruby_gems_client_with_cache'
+require './lib/client/ruby_gems_client'
+require './lib/client/ruby_gems_client_with_cache'
 
 class ShowCommandHandler
   def self.execute(arguments_string)
