+++ lib/models/gem_info.rb
@@ -1,13 +1,10 @@
-require 'pry'
-
 class GemInfo
-  attr_reader :name, :info, :downloads, :licenses, :gem_info
+  attr_reader :name, :info, :downloads, :licenses
 
   def initialize(gem_info)
     @name = gem_info['name']
     @info = gem_info['info']
     @downloads = gem_info['downloads']
     @licenses = gem_info['licenses']
-    @gem_info = gem_info
   end
 end
