+++ lib/printer/result_printer.rb
