+++ spec/client/ruby_gems_client_spec.rb
@@ -1,7 +1,7 @@
-require './lib/ruby_gems_client_with_cache'
+require './lib/client/ruby_gems_client'
 
-RSpec.describe RubyGemsClientWithCache do
-  subject(:client) { RubyGemsClientWithCache }
+RSpec.describe RubyGemsClient do
+  subject(:client) { RubyGemsClient }
 
   describe '.search' do
     subject(:search) { client.search(search_string) }
